+++ RuntimeEditor/EditorIcon.cs
@@ -7,6 +7,9 @@ namespace Hananoki {
 		public static Texture2D Warning => Icon.Get( "console.warnicon.sml" );
 		public static Texture2D Info => Icon.Get( "console.infoicon.sml" );
 		public static Texture2D SceneAsset => Shared.Icon.Get( "$BuildSettings_SelectedIcon" );
+
+		public static Texture2D PlayButton => Icon.Get( "PlayButton" );
+
 		//"console.warnicon.inactive.sml"
 	}
 }